+++ packages/terra-avatar/src/variants/SharedUser.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import styles from '../common/Avatar.module.scss';
 import {
   AVATAR_VARIANTS, setColor,