+++ packages/terra-demographics-banner/src/DemographicsBannerDisplay.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ResponsiveElement from 'terra-responsive-element';
-import 'terra-base/lib/baseStyles';
 
 import './DemographicsBanner.module.scss';
 import SmallDemographicsBannerDisplay from './_SmallDemographicsBannerDisplay';