+++ packages/terra-show-hide/CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Import `injectIntl` and `intlShape` from react-intl instead of terra-base
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 
 2.6.0 - (March 21, 2019)
 ------------------